Abstracts schemas into separate files
