+++ app/schemas/token.py
@@ -1,0 +1,10 @@
+from typing import List, Optional
+from pydantic import BaseModel
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+class TokenPayload(BaseModel):
+    sub: Optional[int] = None
